+++ src/project/api/service.py
@@ -1,13 +1,19 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 
+from project.schemas.user import UserSchema
 from project.schemas.service import ServiceSchema, ServiceCreateUpdateSchema
 from project.core.exceptions.serivce import ServiceNotFound, ServiceNoHotel, ServiceAlreadyExists
-from project.api.depends import database, service_repo
+from project.api.depends import database, service_repo, get_current_user, check_for_admin_access
 
 router = APIRouter()
 
 
-@router.get("/all_services", response_model=list[ServiceSchema], status_code=status.HTTP_200_OK)
+@router.get(
+    "/all_services",
+    response_model=list[ServiceSchema],
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_user)],
+)
 async def get_all_services() -> list[ServiceSchema]:
     async with database.session() as session:
         all_services = await service_repo.get_all_services(session=session)
@@ -15,7 +21,12 @@ async def get_all_services() -> list[ServiceSchema]:
     return all_services
 
 
-@router.get("/service/{service_id}", response_model=ServiceSchema, status_code=status.HTTP_200_OK)
+@router.get(
+    "/service/{service_id}",
+    response_model=ServiceSchema,
+    status_code=status.HTTP_200_OK,
+    dependencies=[Depends(get_current_user)],
+)
 async def get_service_by_id(
     service_id: int,
 ) -> ServiceSchema:
@@ -27,10 +38,16 @@ async def get_service_by_id(
     return service
 
 
-@router.post("/add_service", response_model=ServiceSchema, status_code=status.HTTP_201_CREATED)
+@router.post(
+    "/add_service",
+    response_model=ServiceSchema,
+    status_code=status.HTTP_201_CREATED,
+)
 async def add_service(
     service_dto: ServiceCreateUpdateSchema,
+    current_user: UserSchema = Depends(get_current_user),
 ) -> ServiceSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             new_service = await service_repo.create_service(session=session, service=service_dto)
@@ -49,7 +66,9 @@ async def add_service(
 async def update_service(
     service_id: int,
     service_dto: ServiceCreateUpdateSchema,
+    current_user: UserSchema = Depends(get_current_user),
 ) -> ServiceSchema:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             updated_service = await service_repo.update_service(
@@ -62,10 +81,15 @@ async def update_service(
     return updated_service
 
 
-@router.delete("/delete_service/{service_id}", status_code=status.HTTP_204_NO_CONTENT)
+@router.delete(
+    "/delete_service/{service_id}",
+    status_code=status.HTTP_204_NO_CONTENT,
+)
 async def delete_service(
     service_id: int,
+    current_user: UserSchema = Depends(get_current_user),
 ) -> None:
+    check_for_admin_access(user=current_user)
     try:
         async with database.session() as session:
             service = await service_repo.delete_service(session=session, service_id=service_id)
